adde flask routing for wordlist reader
