+++ Bchain_App/Seedphrase_Generator/Dockerfile
