+++ Bchain_App/Seedphrase_Generator/app.py
