+++ Bchain_App/Seedphrase_Generator/requirements.txt
