+++ Bchain_App/Username_Generator/Application/__init__.py
@@ -1,0 +1,6 @@
+from flask import Flask
+from os import getenv
+
+app = Flask(__name__)
+
+from application import routes
