+++ Bchain_App/WordList_Parser/Dockerfile
@@ -1,0 +1,6 @@
+FROM python:3.6
+COPY requirements.txt . 
+RUN pip3 install -r requirements.txt
+COPY . . 
+EXPOSE 5000
+ENTRYPOINT ["python3","app.py"]
