+++ Bchain_App/WordList_Parser/app.py
@@ -1,0 +1,5 @@
+from flask import Flask
+from application import app
+
+if __name__=='__main__':
+  app.run(host='0.0.0.0', port=5000, debug=True)
